# [FYI] Pulse schedule performance improve