--- a/qiskit/pulse/schedule.py
+++ b/qiskit/pulse/schedule.py
@@ -14,15 +14,15 @@
 
 """Schedule."""
 
-import itertools
 import abc
 from typing import List, Tuple, Iterable, Union, Dict, Callable, Set, Optional, Type
 
-from .timeslots import Interval
 from .channels import Channel
-from .interfaces import ScheduleComponent
-from .timeslots import TimeslotCollection
 from .exceptions import PulseError
+from .interfaces import ScheduleComponent
+from .timeslots import Interval
+from .timeslots import TimeslotCollection
+
 
 # pylint: disable=missing-return-doc
 
@@ -30,21 +30,23 @@ from .exceptions import PulseError
 class Schedule(ScheduleComponent):
     """Schedule of `ScheduleComponent`s. The composite node of a schedule tree."""
     # pylint: disable=missing-type-doc
-    def __init__(self, *schedules: List[Union[ScheduleComponent, Tuple[int, ScheduleComponent]]],
-                 name: Optional[str] = None):
+    def __init__(self, *schedules: Union[ScheduleComponent, Tuple[int, ScheduleComponent]],
+                 name: Optional[str] = None, check_overlap: bool = True):
         """Create empty schedule.
 
         Args:
             *schedules: Child Schedules of this parent Schedule. May either be passed as
                 the list of schedules, or a list of (start_time, schedule) pairs
             name: Name of this schedule
+            check_overlap: True if check overlap of time slots
 
         Raises:
             PulseError: If timeslots intercept.
         """
         self._name = name
         try:
-            timeslots = []
+            self._timeslots = TimeslotCollection()
+
             _children = []
             for sched_pair in schedules:
                 # recreate as sequence starting at 0.
@@ -56,10 +58,11 @@ class Schedule(ScheduleComponent):
                 sched_timeslots = sched.timeslots
                 if insert_time:
                     sched_timeslots = sched_timeslots.shift(insert_time)
-                timeslots.append(sched_timeslots.timeslots)
+                if check_overlap and not self._timeslots.is_mergeable_with(sched_timeslots):
+                    raise PulseError("Cannot create from overlapped schedules")
+                self._timeslots = self._timeslots.merged(sched_timeslots)
                 _children.append(sched_pair)
 
-            self._timeslots = TimeslotCollection(*itertools.chain(*timeslots))
             self.__children = tuple(_children)
             self._buffer = max([child.buffer for _, child in _children]) if _children else 0
 
@@ -91,7 +94,7 @@ class Schedule(ScheduleComponent):
         return self._buffer
 
     @property
-    def channels(self) -> Tuple[Channel]:
+    def channels(self) -> Tuple[Channel, ...]:
         """Returns channels that this schedule uses."""
         return self.timeslots.channels
 
@@ -100,11 +103,11 @@ class Schedule(ScheduleComponent):
         return self.__children
 
     @property
-    def instructions(self) -> Tuple[Tuple[int, 'Instruction']]:
+    def instructions(self) -> Tuple[Tuple[int, 'Instruction'], ...]:
         """Iterable for getting instructions from Schedule tree."""
         return tuple(self._instructions())
 
-    def ch_duration(self, *channels: List[Channel]) -> int:
+    def ch_duration(self, *channels: Channel) -> int:
         """Return duration of schedule over supplied channels.
 
         Args:
@@ -112,7 +115,7 @@ class Schedule(ScheduleComponent):
         """
         return self.timeslots.ch_duration(*channels)
 
-    def ch_start_time(self, *channels: List[Channel]) -> int:
+    def ch_start_time(self, *channels: Channel) -> int:
         """Return minimum start time over supplied channels.
 
         Args:
@@ -120,7 +123,7 @@ class Schedule(ScheduleComponent):
         """
         return self.timeslots.ch_start_time(*channels)
 
-    def ch_stop_time(self, *channels: List[Channel]) -> int:
+    def ch_stop_time(self, *channels: Channel) -> int:
         """Return maximum start time over supplied channels.
 
         Args:
@@ -141,17 +144,19 @@ class Schedule(ScheduleComponent):
         for insert_time, child_sched in self._children:
             yield from child_sched._instructions(time + insert_time)
 
-    def union(self, *schedules: List[ScheduleComponent],
-              name: Optional[str] = None) -> 'Schedule':
+    # pylint: disable=arguments-differ
+    def union(self, *schedules: ScheduleComponent,
+              name: Optional[str] = None, check_overlap: bool = True) -> 'Schedule':
         """Return a new schedule which is the union of `self` and `schedule`.
 
         Args:
             *schedules: Schedules to be take the union with this `Schedule`.
             name: Name of the new schedule. Defaults to name of self
+            check_overlap: True if check overlap of time slots
         """
         if name is None:
             name = self.name
-        return Schedule(self, *schedules, name=name)
+        return Schedule(self, *schedules, name=name, check_overlap=check_overlap)
 
     def shift(self, time: int, name: Optional[str] = None) -> 'Schedule':
         """Return a new schedule shifted forward by `time`.
@@ -162,10 +167,10 @@ class Schedule(ScheduleComponent):
         """
         if name is None:
             name = self.name
-        return Schedule((time, self), name=name)
+        return Schedule((time, self), name=name, check_overlap=False)
 
     def insert(self, start_time: int, schedule: ScheduleComponent, buffer: bool = False,
-               name: Optional[str] = None) -> 'Schedule':
+               name: Optional[str] = None, check_overlap: bool = True) -> 'Schedule':
         """Return a new schedule with `schedule` inserted within `self` at `start_time`.
 
         Args:
@@ -173,10 +178,12 @@ class Schedule(ScheduleComponent):
             schedule: Schedule to insert
             buffer: Whether to obey buffer when inserting
             name: Name of the new schedule. Defaults to name of self
+            check_overlap: True if check overlap of time slots
         """
         if buffer and schedule.buffer and start_time > 0:
             start_time += self.buffer
-        return self.union((start_time, schedule), name=name)
+        return self.union((start_time, schedule), name=name, check_overlap=check_overlap)
+    # pylint: enable=arguments-differ
 
     def append(self, schedule: ScheduleComponent, buffer: bool = True,
                name: Optional[str] = None) -> 'Schedule':
@@ -192,7 +199,7 @@ class Schedule(ScheduleComponent):
         """
         common_channels = set(self.channels) & set(schedule.channels)
         time = self.ch_stop_time(*common_channels)
-        return self.insert(time, schedule, buffer=buffer, name=name)
+        return self.insert(time, schedule, buffer=buffer, name=name, check_overlap=True)
 
     def flatten(self) -> 'Schedule':
         """Return a new schedule which is the flattened schedule contained all `instructions`."""