<!--
⚠️ If you do not respect this template, your pull request will be closed.
⚠️ Your pull request title should be short detailed and understandable for all.
⚠️ Also, please add it in the CHANGELOG file under Unreleased section.
⚠️ If your pull request fixes an open issue, please link to the issue.

✅ I have added the tests to cover my changes.
✅ I have updated the documentation accordingly.
✅ I have read the CONTRIBUTING document.
-->

### Summary
Address #2817. This PR improves the run time of the sample code given in #2817 from 34s (273s) to 0.4s (5.5s) by removing unnecessary time slot overlap check.

### Details and comments
- Deligate overlap check in `TimeslotCollection.merged()` to `is_mergable_with()`
- Introduce option not to check overlap in `insert`, `union`, `__init__` of `Schedule`